aim for better correctness, add rawBody option
So, used with express.bodyParser, this seems to have a weird
ordering/race condition situation. It looks like having two middleware
attempt to parse the req.stream is tricky. Either you set them up in
parallel in a wrapper middleware or do something else. I can't seem to
figure out another option.

the old code relied on signature_calculator getting added to the
middleware chain before express.bodyParser. In this way, it added its
event handlers to the req object and assumed bodyParser would be there
to call hold the middleware chain until 'end' was called on the
request...

I guess another way to run this is to leave the current behavior and
document the dependency on bodyParser or an empty middleware that holds
next() until 'end' is fired... hmm.